backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
skyfield==1.49
httpx==0.27.2
python-dotenv==1.0.1
